Crate noirc_printable_type

Enums

PrintableType

PrintableValue

This is what all formats eventually transform into For example, a toml file will parse into TomlTypes and those TomlTypes will be mapped to Value

PrintableValueDisplay

In order to display a PrintableValue we need a PrintableType to accurately convert the value into a human-readable format.

TryFromParamsError

Functions

convert_fmt_string_inputs 🔒

Decode parameters for a call with format string.

convert_string_inputs 🔒

Decode parameters for a normal call, without format string.

decode_printable_value

Assumes that field_iterator contains enough field elements in order to decode the PrintableType.

decode_string_value

fetch_printable_type 🔒

Decode the JSON type descriptor of the arguments passed to the print.

flatten_inputs 🔒

Flatten input parameters into a field vector.

flattened_reference_size 🔒

Calculate the number of fields used represent a reference to a type.

to_string 🔒

Format a given PrintableValue according to an expected PrintableType.

write_template_replacing_interpolations 🔒